requirements.txt
fastapi
uvicorn
requests
rich
psutil
pyyaml
pydantic
click
asyncio
tqdm
colorama
send2trash
apscheduler
